+++ CHANGELOG.md
@@ -7,6 +7,7 @@ Please not that there was no stable release yet and things may break at any time
 
 * Support upgrade functions. Closes: [#1](https://github.com/nponiros/sync_client/issues/1)
 * Configurable partialsThreshold. Closes: [#2](https://github.com/nponiros/sync_client/issues/2)
+* Add getStatus method to get the status as text
 
 ## [1.0.0-beta.5] - 2017-01-22